start email system tests
add nodemailer smtp server and define routing and returns. add server to
docker compose. add email system test that hits said server.

I am committing this even though it is not done because it at least
passes testing right now. for some reason adding an expect is breaking
the whole this so that's cool.