improve email system tests and refactor to hit index
added logic in the mock smtp server to parse out the attachment and send
the text contained in it back to sender. This makes it much more easy to
validate as we can now explicitly confirm the attachment is on the
email. refactored the email system tests to hit index method rather than
email lib directly