add verification methods to FTP, overhaul most of the ftp code
added the ability to list the file directory for FTP when getting a
file. Chained that into methods that allow us to filter down to the
specific file, hash the metadata, and validate the hash. overhauled a
significant portion of the FTP code; biggest change is in the connection
setup. extracted that to its own method as the shared connection
requires more thoughtful handling. extracted the getting and setting
functions from utilities to index as they were breaking the ability to
reference utilities from the ftp and sftp files. added a system test for GET ftps
that's failing right now. overhauled some sanctuary definintions to be
more specific. added todos in some places. added a bunch of tests for
the new sub-methods in the ftp file