Add recursion to s/ftp file verification
Added recursive functions to both the ftp and sftp files so that we can
verify the file signature as many times as we want. Migrated file
signature out of connection config object. Included F$ in sanctuary
types setup as Either(Future)(Future) was not working without it. Added
tests for new functions. Added no-unused-vars rule to eslint.