Fix FTPS system test error by mandating TLSv1.2
This was a doozy of an issue. Michael and I spent a few hours pouring
through node's TCP docs as well as VSFTPD's docs. VSFTPD does not
currently support TLSv1.3 even though openssl does. Hardcoding the TLS
version in the connectionConfigs fixed it. Not a well documented issue
but https://forum.filezilla-project.org/viewtopic.php?t=50598 summarizes
it fairly well.